requirements.txt
# for AI
python-dotenv
tqdm
openai
mediapipe 
opencv-python

#for web
Flask 
flask-cors 
requests
pillow

#for database
#conda-forge 
pymongo
dnspython 
email-validator
Flask-PyMongo